Remove req for feedback comment and change missing assignment warning
Feedback column previously was a required component of the JSON and if
not present markstopdf.js would complain and exit. For some assignments
a feedback column is not needed, so this was changed to a warning when
this field is ommitted from the json.

Errors were previously printed for each student what was located in the
CSV, but a matching canvas submission could not be found. This resulted
in many false positives, since every student who hadn't submitted was
printed to the console as an 'error'. The logic was changed so that only
if a submission is missing and a student's grade was NON-ZERO in the
CSV (indicating they likely didn't submit), A large warning prints to
the console. If a submission for a student with a 0 grade is not found,
then only a small error is printed.